Minor renamings
