+++ system/src/control/log_config.cpp
@@ -15,7 +15,7 @@
  * License along with this library; if not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "logging.h"
+#include "log_config.h"
 
 #if SYSTEM_CONTROL_ENABLED
